// src/entities/building/model/types.ts
export interface Building {
  id: number;
  name: string;
  icon: string;
  image: string;
  link: string;
  position: {
    pc: { top: string; left?: string; right?: string };
    tablet: { top: string; left?: string; right?: string };
    mobile: { top: string; left?: string; right?: string };
  };
}